ISSUE#21: Browser name find for linux and windows
Browser names need to be converted from the normalized name
given in the command line (e.g. "chrome") to
a name that matches the associated executable on the
operating system (e.g. "Google Chrome" on macos or "google-chrome" on
linux).

This commit adds support of this function for windows and linux
additionally to the existing support for macos (darwin).

Signed-off-by: FrontSide <david@isan.engineer>